Make '+' character part of word/morpheme
This results in many fewer words (93762 instead of 95794), because
words with '+' in non-word-initial positions were incorrectly split,
which is now fixed.
It also results in 8 fewer morphemes (120134 instead of 120141).
That is the result of typos in the source text (e.g., missing
spaces between words or '+' after initial alaph), which combines
combinations that were separated by '+' previously.